Set up arclint and add flow-array-type-style, flow-no-one-tuple rules
Summary:
This diff is for the recently created eslint-plugin-khan repo.  The purpose
of this repo is to house custom eslint rules we want for use a KA.  The
reason this isn't part of khan-linter proper is that it'll be easier for external
people to use this and contribute to this set of eslint rules.

The flow-array-type-style rule accepts Array<T> flow types and
rejects T[] flow types.  The flow-no-one-tuple rule disallows [T] which
is a one-tuple and will autofix it to Array<T> which is probably what
was meant.

https://app.asana.com/0/325405028872522/323663604270817
https://app.asana.com/0/325405028872522/332728652278742

Test Plan: npm test

Reviewers: joshua, mattdunnrankin, jared, john

Reviewed By: joshua, mattdunnrankin, jared

Differential Revision: https://phabricator.khanacademy.org/D35488